HSphaseplot makes a phase plot of predicted mature depletion vs H

HSphaseplot generates, for a single sau, a Kobe or phase plot of the predicted harvest rate against the predicted mature biomass depletion. It highlights the startyr of cpue conditioning (1992 in Tasmania) the end of conditioning (usually 2020 in Tasmania), and the final year (2050 if projected for 30 years). The default limit on harvest rate is set at 0.175, this may not be precautionary for all sau.

HSphaseplot(
  dyn,
  glb,
  sau,
  rundir = "",
  startyr = 1992,
  console = TRUE,
  targdepl = 0.4,
  limdepl = 0.2,
  limH = 0.175
)

Arguments

dyn

a list of the dynamics for every sau. This function only requires the mature biomass depletion and the harvest rates

glb

the globals object

sau

which sau index (1 - 8 in Tasmania) to use?

rundir

the directory for all files and results, default=""

startyr

default = 1992. Otherwise could use condC$yearCE[1]

console

should the plot go to the console or be saved? Default=TRUE

targdepl

what target depletion is wanted, default=0.4

limdepl

what is the limit depletion reference point, default = 0.2

limH

what is the maximum H of limit fishing mortality, default = 0.2 #'

Value

a list of the median mature biomass depletion and median harvest rate by year invisibly, it also plots these vectors.

Examples

print("wait on a data set")
#> [1] "wait on a data set"
# for (plotsau in 1:glb$nSAU) {
#   kobedata[[plotsau]] <- HSphaseplot(dyn=sauout,glb=glb,sau=plotsau,
#                                      rundir=rundir,startyr=condC$yearCE[1],
#                                      console=FALSE,targdepl=kobeRP[1],
#                                      limdepl=kobeRP[2],limH=kobeRP[3])
# }