STM Generates the Size Transition Matrix for Inverse Logistic

STM With the input of the four parameters inside a vector, and a vector of initial lengths or mid-points of size classes STM generates a square transition matrix with the probabilities of growing from each initial size into the same or larger sizes. Negative growth is disallowed. All columns in the matrix sunm to one.

STM(p, mids)

Arguments

p

a vector of four parameters in the following order MaxDL the maximum growth increment of the inverse logistic, L50 the initial length at which the inflexion of the growth increment curve occurs, L95 - the initial length that defines the 95th percentile of the growth increments, SigMax - the maximum standard deviaiton of the normal distribution used to describe the spread of each distribution of growth increments

mids

a vector of initial lengths which also define the width of each size class thus, mids from 2 - 210 woul dimply 2mm size classes 1 - 3 = 2, 3 - 5 = 4, etc

Value

A square matrix with dimension =the length of the mids vector

References

Haddon, M., Mundy, C., and D. Tarbath (2008) Using an inverse-logistic model to describe growth increments of blackip abalone (Haliotis rubra) in Tasmania. Fisheries Bulletin 106: 58-71

Examples

 param <- c(25.0,120.0,170.0,4.0)
 midpts <- seq(2,210,2)
 G <- STM(param,midpts)
 print(round(G[1:30,1:8],4))
#>         [,1]   [,2]   [,3]   [,4]   [,5]   [,6]   [,7]   [,8]
#>  [1,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#>  [2,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#>  [3,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#>  [4,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#>  [5,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#>  [6,] 0.0002 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#>  [7,] 0.0011 0.0002 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#>  [8,] 0.0049 0.0011 0.0002 0.0000 0.0000 0.0000 0.0000 0.0000
#>  [9,] 0.0168 0.0049 0.0011 0.0002 0.0000 0.0000 0.0000 0.0000
#> [10,] 0.0445 0.0168 0.0050 0.0011 0.0002 0.0000 0.0000 0.0000
#> [11,] 0.0925 0.0446 0.0168 0.0050 0.0012 0.0002 0.0000 0.0000
#> [12,] 0.1505 0.0926 0.0446 0.0169 0.0050 0.0012 0.0002 0.0000
#> [13,] 0.1917 0.1506 0.0927 0.0447 0.0169 0.0050 0.0012 0.0002
#> [14,] 0.1912 0.1918 0.1507 0.0928 0.0448 0.0169 0.0050 0.0012
#> [15,] 0.1492 0.1911 0.1918 0.1508 0.0929 0.0449 0.0170 0.0050
#> [16,] 0.0912 0.1491 0.1911 0.1919 0.1509 0.0931 0.0450 0.0170
#> [17,] 0.0436 0.0911 0.1490 0.1911 0.1919 0.1511 0.0932 0.0451
#> [18,] 0.0163 0.0435 0.0910 0.1489 0.1910 0.1920 0.1512 0.0934
#> [19,] 0.0048 0.0163 0.0435 0.0909 0.1488 0.1909 0.1920 0.1514
#> [20,] 0.0011 0.0048 0.0163 0.0434 0.0908 0.1487 0.1909 0.1921
#> [21,] 0.0002 0.0011 0.0048 0.0162 0.0433 0.0906 0.1485 0.1908
#> [22,] 0.0000 0.0002 0.0011 0.0047 0.0162 0.0432 0.0905 0.1484
#> [23,] 0.0000 0.0000 0.0002 0.0011 0.0047 0.0162 0.0431 0.0903
#> [24,] 0.0000 0.0000 0.0000 0.0002 0.0011 0.0047 0.0161 0.0430
#> [25,] 0.0000 0.0000 0.0000 0.0000 0.0002 0.0011 0.0047 0.0160
#> [26,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0002 0.0011 0.0047
#> [27,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0002 0.0011
#> [28,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0002
#> [29,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#> [30,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000