adjustavrec re-fits the AvRec saudatafile values in the MSE

adjustavrec is used during the conditioning of the MSE to improve the fit to the time-series of CPUE by adjusting the SAU initial value for AvRec. Once the different SAU have been characterized and the data file values settled (even when sizemod has been used), when do_condition is run the fit of the predicted cpue and size-composition data can often be improved by re-fitting the AvRec constants in each SAU to adjust their values to better suit the population characteristics once random variation is added to each population's properties. adjustavrec does this automatically. Be wary of this function because it directly alters the values of AvRec inside the saudata.csv file. If the differences remain too great it may be necessary to adhust the recdevs as well (see optimizerecdevs).

adjustavrec(
  rundir,
  glb,
  ctrl,
  calcpopC,
  verbose = TRUE,
  lowmult = 0.6,
  highmult = 1.4,
  iterlim = 30
)

Arguments

rundir

the full path to the directory in which all files relating to a particular run are to be held.

glb

the globals object

ctrl

the control object - contains datafile and controlfile names

calcpopC

a function that takes the output from hcrfun and generates the actual catch per population expected in the current year.

verbose

Should progress comments be printed to console, default=TRUE

lowmult

multiplier to determine low range of search, default=0.6

highmult

multiplier to determine the high range of search, default=1.4

iterlim

the maximum number of iterations in the search for a solution, default=30

Value

nothing though it will write results to the console if verbose=TRUE

See also

Examples

if (FALSE) {
  # needs at least libraries, aMSE, hutils, makehtml, TasHS
  dropdir <- getDBdir()
  prefixdir <- paste0(dropdir,"A_codeUse/aMSEUse/scenarios/")
  startime <- Sys.time()
  postfixdir <- "S21"
  verbose <- TRUE
  rundir <- filenametopath(prefixdir,postfixdir)
  controlfile <- paste0("control",postfixdir,".csv")
  confirmdir(rundir)
  data(zone1)
  data(saudat)
  rewritecontrolfile(rundir,zone1,controlfile=controlfile)
  rewritedatafile(rundir,zone1,saudat)
  rewritecompdata(rundir,zone1)
  hsargs <- list(mult=0.1,wid = 4,targqnt = 0.55,
                 maxtarg = c(150,150,150,150,150,150,150,150),
                 pmwts = c(0.65,0.25,0.1),
                 hcr = c(0.25,0.75,0.8,0.85,0.9,1,1.05,1.1,1.15,1.2),
                 startCE = 1992)
  out <- do_condition(rundir,controlfile,calcpopC=calcexpectpopC,
                      verbose = TRUE,doproduct = TRUE,dohistoric=TRUE,
                      mincount=120)
  makeoutput(out,rundir,postfixdir,controlfile,hsfile="TasHS Package",
             doproject=FALSE,openfile=TRUE,verbose=FALSE)
  adjustavrec(rundir,out$glb,out$ctrl,calcpopC=calcexpectpopC,verbose=TRUE)
 # Now repeat the do_condition and makeoutput steps to see the final result
}