calcprojsel generates the selectivity and selectivity x weight

calcprojsel is used to produce each projection years' selectivity and selectivity x weight-at-size. projSel is an Nclass x projyrs matrix, a separate selectivity for each projection year, which allows for changing the selectivity during the projection. The projSelWt is a Nclass x projyrs x numpop array, required because each population's weight-at-size relationship will be slightly different. Now appears to be redundant.

calcprojsel(zoneC, projC, glb)

Arguments

zoneC

the constants object

projC

the projection object from zone1

glb

the globals object

Value

a list of projSel and projSelWt

Examples

 print("wait on new example data")
#> [1] "wait on new example data"