changeline replaces a given line in a given file with new text

changeline enables a text file to be changed line by line. One identifies a given line, perhaps by using 'findlinenumber', and this can then be replaced by an input character string. Obviously this is a fine way to mess up a data file so use with care. Used in optimizerecdevs and sauavrecssq

changeline(indir, filename, linenumber, newline, verbose = FALSE)

Arguments

indir

the directory path in which to find the text file. Usually, rundir

filename

the full name of the text file in quotations.

linenumber

either the line number within the text file to be changed, or, the character name of the variable to be changed, e.g. 'AvRec'

newline

the character string with which to replace the line

verbose

should confirmation be output to the console. default=FALSE

Value

nothing but it does alter a line in a text file. Optionally it may confirm the action to the console

See also

Examples

print("wait on an example")
#> [1] "wait on an example"