do_condition is a function to assist with conditioning the model

do_condition is a utility function that can be used to speed the conditioning of the operating model on available fishery data. During the generation of the simulated zone the operation that takes the longest is the estimation of the productivity of each population. The productivity is estimated at equilibrium and getting there is what takes the time. When conditioning the model the objective is to match the predicted fishery response to the historical catches to the observed responses. 'do_condition' generates the simulated zone without running the projections, which simplifies any searches for optimal values of AvRec (average unfished recruitment), and especially for suitable values of recruitment deviates. If any of the initdepl (initial depletion) values, as listed in the control file are less than 1.0 then before applying the historical catches it first depletes each population within each SAU to the initdepl value for each SAU. For this to work doproduct must be TRUE. This will slow down any searches for an optimum set of AvRec and recdevs, so a different strategy for doing that will be needed.

do_condition(
  rundir,
  controlfile,
  calcpopC,
  verbose = FALSE,
  doproduct = FALSE,
  dohistoric = TRUE,
  matureL = c(70, 200),
  wtatL = c(80, 200),
  mincount = 100
)

Arguments

rundir

the full path to the directory in which all files relating to a particular run are to be held.

controlfile

the filename of the control file present in rundir containing information regarding the particular run.

calcpopC

a function that takes the output from hcrfun (either the aspirational catch x SAU or TAC x zone) and generates the actual catch per population in each SAU expected in the current year.

verbose

Should progress comments be printed to console, default=FALSE

doproduct

should production estimates be made. default=FALSE

dohistoric

should the historical catches be applied. Default=TRUE

matureL

is a vector of 2, default = c(70,200), that define the x-axis bounds on the biology maturity-at-Length plots

wtatL

is a vector of 2, default = c(80,200), that define the x-axis bounds on the biology weight-at-Length plots

mincount

determines the minimum sample size for a size-composition sample to be included in plots and analyses. Default = 100

Value

a large list containing tottime, projtime, starttime, glb, ctrl, zoneDD, zoneDP, projC, condC, sauout, and outzone

See also

Examples

print("wait on suitable data sets in data")
#> [1] "wait on suitable data sets in data"
# rundir=rundir; controlfile=controlfile;calcpopC=calcexpectpopC
# verbose=TRUE; doproduct=TRUE; dohistoric=TRUE; mincount=120
# matureL=c(70,200);wtatL=c(80,200);mincount=120