dohistoricC imposes the historical catches on an unfished zone

dohistoricC is used during the conditioning of the zone/region and imposes the historical catches, held in the zone data object obtained using readzonefile, onto an unfished initial zone, and it does this by imposing the time-series of catches to each SAU/block. The operation is through the use of oneyearC, which imposes one year's catches, which are a vector of SAU catches for each year of the series.

dohistoricC(
  zoneDD,
  zoneC,
  glob,
  condC,
  calcpopC,
  fleetdyn,
  hsargs,
  sigR = 1e-08,
  sigB = 1e-08
)

Arguments

zoneDD

The input unfished dynamic zone, zoneD, object

zoneC

the zone constants object, zoneC

glob

the globals object

condC

from the zone1 object, contains historical fisheries data plus

calcpopC

a function that takes the output from hcrfun and generates the actual catch per population expected in the current year.

fleetdyn

a function that calculates the distribution of catch across the sau and populations. Currently not needed by Tas but needed by SA

hsargs

the constants used to define the workings of the hcr

sigR

the recruitment variation included

sigB

the variability introduced to the catches by population by fishers not knowing the distribution of exploitable biomass exactly. What this value should be is unknown, the default=1e-08, is arbitrary but avoids any effective fisher allocation error between populations.

Value

a zoneD object

See also

Examples

print("wait on some data sets")
#> [1] "wait on some data sets"
#  zoneD,zoneC,glob=glb,condC,calcpopC=calcpopC,sigR=1e-08,sigB=1e-08