doprojections conducts the replicate model runs for Tasmania

doprojections conducts the replicate model runs for Tasmania using the mcdahcr and the input hsargs. Note the use of a list that takes up both the population and catch numbers-at-size and puts that into getdata. By referencing the iteration in the arrays being passed to getdata the object sizes are being reduced and hence, hopefully, the runtime will be reduced accordingly.

doprojections(
  ctrl,
  zoneDP,
  zoneCP,
  glb,
  hcrfun,
  hsargs,
  sampleCE,
  sampleFIS,
  sampleNaS,
  getdata,
  calcpopC,
  makehcrout,
  fleetdyn,
  verbose = FALSE,
  ...
)

Arguments

ctrl

the ctrl object from readctrlfile

zoneDP

the object used to contain the dynamics from the replicate model runs

zoneCP

the object used to contain the constants for each population used in model dynamics

glb

the object containing the global constants for the given run

hcrfun

the name of the harvest control rule that is used to calculate the multiplier for the previous aspirational catches (possibly for each SAU but possibly the TAC for the whole zone) so as to estimate the aspirational catches or TAC or the following year

hsargs

the constants used to define the workings of the hcr

sampleCE

a function that generates the CPUE statistics

sampleFIS

a function that generates the FIS statistics

sampleNaS

a function that generates the Numbers-at-size samples

getdata

a function that gathers all the data required by the hcrfun and combines it into an hcrdata object ready for the hcrfun. It is expected to call sampleCE, sampleFIS, and sampleNAS, even if they only return NULL.

calcpopC

a function that takes the output from hcrfun and generates the actual catch per population expected in the current year.

makehcrout

is a function from HS.R that produces an object that is updated in each iteration by the hcrfun. If no such object is required then have a function that returns NULL.

fleetdyn

a function that calculates the distribution of catch across the sau and populations. Currently not needed by Tas but needed by SA

verbose

should the iterations be counted on the console?

...

the ellipsis used in case any of the functions hcrfun, sampleCE, sampleFIS, sampleNas, and getdata require extra arguments not included in the default named collection

Value

a list containing the full dynamics across all years zoneDP, and the final output from the HCR as hcrout

See also

Examples

print("wait on suitable internal data sets")
#> [1] "wait on suitable internal data sets"