doquantribbon generates a ribbon of multiple quantile values

doquantribbon is used when trying to present a summary of multiple scenarios applied across a sequence of years (for example in the projection years of an MSE). It assumes that for each scenario quantiles have been calculated for each year so that the input array if 5 x nyrs x nscenes, where nyrs is the number of years and nscenes is the number of scenarios. Five quantiles are assumed with values probs=c(0.025, 0.05, 0.5, 0.95 0.975), and the option exists of plotting either the 90th (0.05 and 0.95) or the 95th (0.025, 0.975). If the input quantiles have different values then the 'q90' argument,if TRUE, will reference the 2nd and 4th row and the 1st and 5th if FALSE. If polys is TRUE then filled transparent polygons will be plotted, with colours in the same sequence as the scenarios, alternatively, if polys=FALSE, then lines will be plotted instead. This function is currently only used by sauribbon.

doquantribbon(
  varq,
  varname,
  yrnames,
  scenes,
  q90 = TRUE,
  intens = 127,
  addleg = "bottomright"
)

Arguments

varq

an array of quantiles with dimensions the 5 quantiles of the variable in question (2.5, 5, 50, 95 97.5), by years, the number of years for which quantiles are available

varname

The names of the variable being summarized, used as the y-label on the plot

yrnames

the numeric values of the years to be plotted. Used as the x-axis as well as the x-axis labels

scenes

the names given to the different scenarios

q90

should the 90th or 95th quantiles be plotted, default=TRUE = 90th

intens

signifies the intensity of colour on a scale of 0 - 255. 127 is about 50 percent dense.

addleg

add a legend? currently can only use topleft, topright, bottomleft, or bottomright = default

Value

nothing but it does add polygons to a plot

See also

Examples

print("wait on datasets")
#> [1] "wait on datasets"