dosauplot generates the plot of the chosen variable for each sau

dosauplot takes the results of the prepareproject and the projection function, it combines the trajectories across the years of conditioning and the projection years, for all replicates, and plots all replicates for the years from 'startyr' to the final year of the projections. It includes the median and inner 90 plot, if the historical conditioning CPUE is included in histCE, it also includes the original cpue series. Beware of trying to plot the historical cpue in years where there are no historical cpue data, although there are error captures at work to limit the pot to available.

dosauplot(
  ylabel,
  postrep,
  glb,
  startyr,
  addCI = FALSE,
  CIprobs = c(0.05, 0.5, 0.95),
  histCE = NULL,
  addCE = FALSE
)

Arguments

ylabel

the variable name to be plotted, used as a y-axis label

postrep

the replicate values from the replicated projection zoneDP for the variable selected taken from the output of 'zonetosau'.

glb

the global constants object

startyr

the index for the first year of the conditioning dynamics to include

addCI

should quantile confidence bounds be included on the plots, default=FALSE

CIprobs

the quantiles used to generate the confidence intervals. The default = c(0.05,0.5,0.95)

histCE

the historical cpue data if included, default = NULL

addCE

should historical cpue be included on the plots, default=FALSE

Value

invisibly, a list of CI and median for each SAU

Examples

print("wait on suitable built in data sets")
#> [1] "wait on suitable built in data sets"