findlinenumber prints out the contents of a text file with line numbers

findlinenumber solves the problem of finding the line number in a given text file when one wants to change a specific line using the function changeline. After inc lines are printed the function stops and waits for any character to be input (a space will suffice) before printing the next inc lines

findlinenumber(rundir, filename, inc = 20)

Arguments

rundir

the directory path in which to find the text file.

filename

the full name of the text file in quotations.

inc

the number of lines on screen

Value

nothing, but it does print the contents of a text file with the respective line numbers on the console.

Examples

print("wait on a suitable example")
#> [1] "wait on a suitable example"