getlistvar extracts a vector or matrix from zoneC

getlistvar extracts a vector or matrix from zoneC. If a vector of scalars, the names relate to populations, if a matrix the columns relate to populations. Only Me, R0, B0, effB0, ExB0, effExB0, MSY, MSYDepl, bLML, scalece, SaM, SAU, popdef, LML, Maturity, WtL, Emergent, and MatWt are currently valid choices. The indexvar = popdef would generate a listing of all the constants. If you only want a single constant from popdefs then use indexvar2.

getlistvar(zoneC, indexvar, indexvar2 = "")

Arguments

zoneC

the constants components of the simulated zone

indexvar

the name of the variable to be extracted; character

indexvar2

the name of the variable within popdef to extract

Value

either a vector or matrix of values depending on variable

Examples

 data(zone)
 zoneC <- zone$zoneC
 getlistvar(zoneC,"MSY")
#>       p1       p2 
#> 207.5093 207.5093 
 getlistvar(zoneC,"B0")
#>       p1       p2 
#> 4148.483 4148.483 
 getlistvar(zoneC,"popdef","AvRec")
#>      p1      p2 
#> 1764761 1764761