getssqparts calculates the ssq for the cpue and size-comp data

getssqparts is used during the conditioning of the operating model to improve the fit of the model parameters to the observed SAU scale data on cpue and size-composition. It only uses sum-of squared deviations so an arbitrary weight is given to the size-composition data, and that value should be such that changes in the ssq for the cpue are of the same order of magnitude for the size-composition data.

getssqparts(rundir, controlfile, calcpopC, mincount = 100, wtsc = 0.02)

Arguments

rundir

the directory in which all the files and outputs are kept

controlfile

the name of the controlfile for the MSE run

calcpopC

the function from the harvest strategy that is used to calculate the expected catch by population within each say each year. The annual catch by sau is known but in the MSE that needs to be distributed, in this case without error, across each population.

mincount

the minimum count used in the size-composition data selection default = 100, this can be a vector of length nsau with a value for each sau if you want to give a different value to each sau.

wtsc

the weighting given to the size-composition data ssq, default= 0.02. This can be a vector of length nsau if you want to give each sau a different weighting.

Value

a vector of the total ssq, the cpue ssq, and the size-comp ssq

Examples

print("wait on internal data sets")
#> [1] "wait on internal data sets"