getzoneLF extracts all LF data from a zone across pops and years

getzoneLF extracts all LF data from a zone across all populations for each year. Thus an Nclass x Nyrs X numpop matrix is compressed into a Nclass x Nyrs matrix

getzoneLF(zoneD, glb)

Arguments

zoneD

the dynamic part of the zone

glb

the global constants

Value

an Nclass x Nyrs matrix containing LF data across all populations by year

Examples

data(zone)
zonelf <- getzoneLF(zone$zoneD,zone$glb)
str(zonelf)
#>  num [1:105, 1:58] 3.53e+06 5.43e-08 4.34e-06 2.46e-04 9.89e-03 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : chr [1:105] "2" "4" "6" "8" ...
#>   ..$ : chr [1:58] "1" "2" "3" "4" ...