getzoneprod zone scale summary of product matrix from doproduction

getzoneprod takes in the product matrix from doproduction and sums the mature and exploitable biomassand the total catches. Then it recalculates, for the zone, the approximate annual harvest rate, the depletion level, and the relative catch rate. The annual harvest rate and relative catch rate are simply the arithmetic mean of the values for all populations, which the depletion is the column of mature biomass divided by the first value = B0

getzoneprod(product)

Arguments

product

The productivity array from doproduction containing the range of imposed harvest rates, and the resulting outputs for each population

Value

a matrix containing the approximate productivity matrix for the zone

Examples

data(zone)
outprod <- getzoneprod(zone$product)
print(outprod[1:25,])
#>            ExB     MatB  AnnH     Catch    Deplet    RelCE
#> 0     7598.331 8296.967 0.000   0.00000 1.0000000 376.4738
#> 0.005 7245.699 7934.393 0.005  37.67598 0.9563004 363.2924
#> 0.01  6916.797 7596.178 0.010  72.10314 0.9155368 350.8527
#> 0.015 6609.434 7280.065 0.015 103.59555 0.8774370 339.0933
#> 0.02  6321.676 6984.059 0.020 132.42976 0.8417606 327.9595
#> 0.025 6051.807 6706.393 0.025 158.85019 0.8082946 317.4021
#> 0.03  5798.303 6445.497 0.030 183.07352 0.7768498 307.3773
#> 0.035 5559.805 6199.969 0.035 205.29247 0.7472573 297.8453
#> 0.04  5335.095 5968.557 0.040 225.67897 0.7193662 288.7704
#> 0.045 5123.083 5750.139 0.045 244.38680 0.6930411 280.1203
#> 0.05  4922.787 5543.703 0.050 261.55389 0.6681602 271.8656
#> 0.055 4733.322 5348.339 0.055 277.30427 0.6446138 263.9797
#> 0.06  4553.888 5163.225 0.060 291.74973 0.6223027 256.4382
#> 0.065 4383.760 4987.614 0.065 304.99128 0.6011371 249.2190
#> 0.07  4222.280 4820.830 0.070 317.12037 0.5810353 242.3016
#> 0.075 4068.849 4662.257 0.075 328.21997 0.5619231 235.6675
#> 0.08  3922.920 4511.335 0.080 338.36553 0.5437330 229.2995
#> 0.085 3783.995 4367.549 0.085 347.62577 0.5264031 223.1819
#> 0.09  3651.617 4230.432 0.090 356.06339 0.5098769 217.3001
#> 0.095 3525.367 4099.553 0.095 363.73574 0.4941026 211.6408
#> 0.1   3404.861 3974.517 0.100 370.69530 0.4790325 206.1914
#> 0.105 3289.744 3854.961 0.105 376.99024 0.4646229 200.9406
#> 0.11  3179.690 3740.551 0.110 382.66480 0.4508335 195.8775
#> 0.115 3074.398 3630.977 0.115 387.75965 0.4376271 190.9924
#> 0.12  2973.591 3525.955 0.120 392.31232 0.4249692 186.2760