lfs is a 3-D array of size-composition data for use in examples

lfs is a 38 x 31 x 8 array of values for size-composition data from the west coast of Tasmania made up of 38 size classes, 31 years, from 1990 - 2020, and 8 statistical blocks. This is purely for use with examples within the aMSE R package and requires quality control checks before use anywhere else.

Examples

 data(lfs)
 lfs[,10:20,6]
#>     1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009
#> 136    0    0    0    0    0    0    0    0    0    5    5
#> 138    0    1    0    1    1    1    0    0    2    7    3
#> 140    0    4    4    9    3   10    2    1   11   30   16
#> 142    0    8    5   15   11   17    3    1   30   66   28
#> 144    0    5   21    6   14   19    7    4   39   64   31
#> 146    0   11   18   11   17   27   11   11   44   92   48
#> 148    0    7   17   19   17   15   17   10   47  105   36
#> 150    0   17   23   14   17   24    6   16   62  124   37
#> 152    0   10   21   20   22   25   15   11   57  112   33
#> 154    0   16   23   15   18   26   19   17   63  113   47
#> 156    0   14   21   11   27   20   25   10   54  113   41
#> 158    0   10   21   11   16   25   13   21   62  119   43
#> 160    0    5   23    9   13   17   23   11   58  108   27
#> 162    0   13   19   14   14   27   22   20   43   91   24
#> 164    0    4   22   14   22    8   14   16   55   76   25
#> 166    0    4   16    8   17    9   19   10   47   65   19
#> 168    0    3   13   13   17    5   21   12   27   46   10
#> 170    0    3    9    1   20    3   23    4   24   42    8
#> 172    0    6    6    8   14    3    8    4   16   24    7
#> 174    0    3    9    2    9   11   11    8   12   27    2
#> 176    0    1    8    4    2    0   13    4    6   17    1
#> 178    0    1    2    6    6    4   11    3    3   15    4
#> 180    0    4    6    4    1    1    1    1    6    9    3
#> 182    0    0    2    2    1    3    9    1    3    4    1
#> 184    0    0    0    0    1    0    3    2    0    2    1
#> 186    0    0    0    0    0    0    4    1    0    4    0
#> 188    0    0    1    1    0    0    0    2    0    0    0
#> 190    0    1    0    0    0    0    2    0    1    3    1
#> 192    0    0    0    0    0    0    1    0    0    0    0
#> 194    0    0    0    0    0    0    0    0    0    1    0
#> 196    0    0    0    0    0    0    0    0    0    0    0
#> 198    0    0    0    0    0    0    0    0    1    0    0
#> 200    0    0    0    0    0    0    0    0    0    0    0
#> 202    0    0    0    0    0    0    0    0    0    0    0
#> 204    0    0    0    0    0    0    0    0    0    0    0
#> 206    0    0    0    0    0    0    0    0    0    0    0
#> 208    0    0    0    0    0    0    0    0    0    0    0
#> 210    0    0    0    0    0    0    0    0    0    0    0