makeequilzone high level function generates equilibrium zone

makeequilzone is a high level function that merely hides the details of generating the original unfished zone after reading in the data files, estimates the productivity, and sets up the results directory, rundir, ready to receive files.

makeequilzone(
  rundir,
  ctrlfile = "control.csv",
  doproduct = TRUE,
  verbose = TRUE
)

Arguments

rundir

the directory containing the control csv files. It can/will also act to store results in a manner that will allow them to be displayed using makehtml.

ctrlfile

the main file that controls the particular run. It contains the name of the data file that is used to biologically condition the numpop populations

doproduct

boolean, should the productivity calculations be made during the conditioning. Set to FALSE conditionOM

verbose

Should progress comments be printed to console, default=TRUE

Value

a list of zoneC, zoneD, glb, constants, saudat,product, ctrl, and zone1

Examples

print("wait on datafiles")
#> [1] "wait on datafiles"