makeoutput take the output from do_MSE and generates the HTML files

makeoutput simplifies taking the output from do_MSE and producing the HTML files used to display all the results in rundir. Note that 'runnotes' is now a vector of character strings made up using paste0.

makeoutput(
  out,
  rundir,
  postdir,
  controlfile,
  hsfile = NULL,
  doproject = TRUE,
  openfile = TRUE,
  verbose = FALSE
)

Arguments

out

the output from do_MSE

rundir

the full path to the directory holding the all scenario files

postdir

the name of the directory holding the results, also used to name the internal webpage

controlfile

the controlfile used to run the MSE

hsfile

the name of the harvest strategy file, default=NULL

doproject

have the projections been run? default = TRUE.

openfile

should the website be opened automatically? default=TRUE

verbose

should details of producing the HTML files be written to the console? default=FALSE

Value

nothing but it does generate a set of HTML files in rundir. If verbose=TRUE it also writes text to the console

Examples

print("wait on internal data-sets")
#> [1] "wait on internal data-sets"
# out=out;rundir=rundir;postdir=postfixdir;controlfile=controlfile;hsfile="TasHS Package"
# doproject=FALSE;openfile=TRUE;verbose=FALSE