makewidedat converts long data to a wide data format

makewidedat takes the output of the commlf function makelongdat and converts the columns of year x length x count into a matrix of counts with axes of lengths x years, ready for comparisons with predicted length composition of catches from the operating model.

makewidedat(inlong, mids, counts = FALSE)

Arguments

inlong

the long form data.frame derived for commlf's makelongdat containing at least 'year', 'length', and 'propcounts'

mids

the range of lengths to be found within the input data file. These are matched with the size distribution for each year of data and the appropriate matrix cells filled with the propcounts.

counts

default = FALSE, which means that the wide format will contain the proportion of counts for each year. If TRUE, then the actual counts by size-class will be output by year

Value

a matrix of proportional counts for each size-class x year

Examples

print("wait on data sets")
#> [1] "wait on data sets"