modzoneC runs the zone 3 x hyrs to equilibrium

modzoneC runs the doproduction function so it can then appropriately fill in each population's MSY and MSYDepl values. It outputs both the finalized zoneC and the prodution array. Hence we need to push copies of both zoneC and zoneD into modzoneC

modzoneC(zoneC, zoneD, glob, lowlim = 0, uplim = 0.4, inc = 0.005)

Arguments

zoneC

the constant components of the simulated zone

zoneD

the dynamic components of the simulated zone

glob

the general global variables

lowlim

the lower limit of harvest rate applied, default=0.0

uplim

the upper limit of harvest rate applied, default=0.4

inc

the harvest rate increment at each step, default=0.005

Value

a list containing the updated zoneC and product

Examples

if (FALSE)   # the doproduction part takes too long to run
  data(zone1)
  glb <- zone1$globals
#> Error in eval(expr, envir, enclos): object 'zone1' not found
  data(constants)
#> Warning: data set 'constants' not found
  ans <- makezoneC(zone1,constants)
#> Error in eval(expr, envir, enclos): object 'zone1' not found
  zoneC <- ans$zoneC
#> Error in eval(expr, envir, enclos): object 'ans' not found
  ans <- makezone(glb,zoneC)
#> Error in eval(expr, envir, enclos): object 'glb' not found
  zoneC <- ans$zoneC
#> Error in eval(expr, envir, enclos): object 'ans' not found
  zoneD <- ans$zoneD
#> Error in eval(expr, envir, enclos): object 'ans' not found
  ans2 <- modzoneC(zoneC,zoneD,glb)
#> Error in eval(expr, envir, enclos): object 'glb' not found
  str(ans2,max.level=2)
#> Error in eval(expr, envir, enclos): object 'ans2' not found
   # zoneC=zoneC; zoneD=zoneD; glob=glb; lowlim=0.0;uplim=0.4;inc=0.01