oneyear one year's harvest dynamics for one abpop

oneyear do one year's dynamics for one input population. Used to step through populations of a zone. Its dynamics are such that it first undergoes growth, then half natural mortality. This allows an estimate of exploitable biomass before fishing occurs. The remaining dynamics involve the removal of the catch, the application of the last half of natural mortality and the addition of recruits. Which allows the exploitable biomass to be estimated after fishing. The recruitment occurs in oneyearD so that larval dispersal can be accounted for. Thus, oneyear requires oneyearD.

oneyear(MatWt, SelWt, selyr, Me, G, qest, WtL, inNt, inH, lambda, scalece)

Arguments

MatWt

maturity x Weight-at-size from zoneCC for the population and year

SelWt

selectivity x Weight-at-size from zoneCC for the population and year

selyr

selectivity from zoneC for the population and year

Me

natural morality from zoneC for the population

G

growth transtion matrix from zoneC for the population

qest

the catchability from zoneCC for the population

WtL

the weight-at-alength from zoneC for the population

inNt

the numbers at size for the year previous to the year of dynamics. These are projected into the active year.

inH

a literal annual harvest rate as a proportion to be removed as catch during the year, a scalar

lambda

the non-linearity parameter. if = 1.0 then cpue linearly related to expB, if < 1 then hyper-stability of cpue occurs.

scalece

the xoneC'pop'$scalece used to scale al cpue to the sau nominla scale

Value

a list containing ExploitB, MatureB, Catch, Harvest, Nt, ce, CatchN, and midyexpB used to update the given pop in yr + 1

Examples

print("need to wait on built in data sets")
#> [1] "need to wait on built in data sets"
#oneyear <- function(MatWt,SelWt,selyr,Me,G,qest,WtL,inNt,inH,lambda,scalece)