oneyearD conducts one year's dynamics on zoneD in the MSE

oneyearD conducts one year's dynamics on zoneD in the MSE returning the revised zoneD, which will have had a single year of activity included in each of its components. This uses zoneC but always within the environment of another function in which zoneC (as zoneC) can be found. Used in runthreeH, and hence in doproduction, and in testequil.

oneyearD(zoneC, zoneD, inHt, year, sigmar, Ncl, npop, movem)

Arguments

zoneC

the constant portion of the zone with a list of properties for each population

zoneD

the dynamics portion of the zone, with matrices and arrays for the dynamic variables of the dynamics of the operating model

inHt

a vector of harvest rates taken in the year from each population

year

the year of the dynamics, would start in year 2 as year 1 is the year of initiation.

sigmar

the variation in recruitment dynamics, set to 1e-08 when searching for equilibria.

Ncl

the number of size classes used to describe size

npop

the number of populations, the global numpop

movem

the larval dispersal movement matrix

Value

a list containing a revised dynamics list

Examples

 print("wait on revised data")
#> [1] "wait on revised data"