oneyearcat one year's catch dynamics for one abpop

oneyear does one year's dynamics for one input population. Where the fishing is based on a given catch per population, which would be determined by any harvest control rule. It is used to step through populations of a zone. Its dynamics are such that each population first undergoes growth, then half natural mortality. This allows an estimate of exploitable biomass before fishing occurs. The remaining dynamics involve the removal of the catch, after estimating the harvest rate from catch/exploitb, the application of the last half of natural mortality and the addition of recruits. This allows the exploitable biomass to be estimated after fishing. The cpue uses the average of the pre-fishing and post-fishing exploitB to smooth over any changes brought about by fishing. The recruitment occurs in oneyearC so that larval dispersal can be accounted for. oneyearcat is not used independently of oneyearC.

oneyearcat(
  MatWt,
  SelWt,
  selyr,
  Me,
  G,
  scalece,
  WtL,
  inNt,
  incat,
  sigce,
  lambda,
  qest,
  fissettings = NULL,
  fisindex = NULL
)

Arguments

MatWt

maturity x Weight-at-size from zoneCC for the population and year

SelWt

selectivity x Weight-at-size from zoneCC for the population and year

selyr

selectivity from zoneCC for the population and year

Me

natural morality from zoneCC for the population

G

growth transtion matrix from zoneCC for the population

scalece

the xoneC'pop'$scalece used to scale al cpue to the sau nominla scale

WtL

the weight-at-alength from zoneCC for the population

inNt

the numbers at size for the year previous to the year of dynamics. These are projected into the active year.

incat

the literal annual catch from the population. Derived from the harvest control rule for each SAU, then allocated to each population with respect to its relative catching performance.

sigce

the process error variation associated with the estimates of cpue from the model. Found in the ctrl object as withsigCE. It is implemented as Log-Normal error on the exploitable biomass value from the model

lambda

the hyper-stability term from zoneC

qest

the estimated catchability from sizemod from zoneC

fissettings

an object containing settings used when calculating indices for the FIS

fisindex

a function used to estimate the FIS index

Value

a list containing a vector of ExploitB, MatureB, Catch, and ce, and a matrix NaL containing Nt and CatchN used to update a pop in yr + 1

See also

Examples

print("need to wait on built in data sets")
#> [1] "need to wait on built in data sets"