oneyearrec calculates the Beverton-Holt recruitment

oneyearrec calculates the Beverton-Holt recruitment for the input populations in a single year; parameterized with steepness, R0, and B0. To drop variation to insignificant levels set sigmar-1-08. To allow for inclusion of recruitment deviates, which can be required to explain high catches when stock biomass would otherwise be low, we have introduced the devR argument. As the deviates are all multiplicative and range in value > 0, then the default = -1 so this will be ignored unless a deviate is input.

oneyearrec(steep, R0, B0, Bsp, sigR, devR = -1, depensate = 0)

Arguments

steep

the steepness for the population; scalar

R0

the unfished recruitment levels for the population; scalar

B0

the unfished spawning biomass; scalar

Bsp

the current spawning biomass; scalar

sigR

standard deviation of the recruitment residuals; scalar. set this to 1e-08 to avoid recruitment variability

devR

if recruitment deviates are available then they should be input here. If negative values then a random epsilon is used, otherwise epsilon is given the value of devR. default=-1, so by default fixed recruitment deviates are off.

depensate

should depensation occur at low biomass levels? default=0, which means no depensation. If one puts, say, 0.2, here then biomass levels from 0.2B0 will lead to a linear decline in recruitment down to zero rather than the B-H curve that normally occurs. any recruitment deviates continue to be applied. Conditioning the model would need to be done with depesation turned on in sizemod as well.

Value

an absolute number of recruits from a given spawning biomass

See also

Examples

if (FALSE) {
steep <- 0.707
R0 <- 319971
B0 <- 313
Bsp <- 147
insigmar <- 0.3
oneyearrec(steep,R0,B0,Bsp,insigmar)
}
# steep=steep;R0=r0;B0=b0;Bsp=matb;sigR=sigmar;devR=-1