oneyearsauC conducts one year's dynamics using catch not harvest

oneyearsauC conducts one year's dynamics in the simulation using historical SAU catches rather than harvest rates. The harvest rates are estimated after first estimating the exploitable biomass. returning the revised zoneD, which will have had a single year of activity included in each of its components. uses the function imperr to introduce implementation error to the actual catches. The option now exists of including one or more years of pperturbation in the survivorship of the recruitment level and post-larval numbers. This is to permit heat wave and similar mortality events impacting on the subsequent recruitment and immediate settled survivorship. To be active envyr should be a vector or which projeciton eyars an impact will occur, and envsurv will be a matrix with a row for each year with the proportion of expected recruitment off the predicted stock recruitment relationship, and the proportion survivorship of post-larval numbers-at-size).

oneyearsauC(
  zoneCC,
  inN,
  popC,
  year,
  Ncl,
  sauindex,
  movem,
  sigmar,
  sigce = 1e-08,
  r0,
  b0,
  exb0,
  rdev = -1,
  envyr,
  envsurv,
  envrec
)

Arguments

zoneCC

the constant portion of the zone with a list of properties for each population ##param exb the vector of exploitable biomass for each population for the ## previous year

inN

the numbers-at-length for each population for the previous year

popC

a vector of actual catches to be taken in the year from each SAU

year

the year of the dynamics the exb and inN would be from year-1

Ncl

the number of size classes used to describe size, global Nclass

sauindex

the sau index for each population from glb

movem

the larval dispersal movement matrix, global move

sigmar

the variation in recruitment dynamics, set to 1e-08 when searching for an equilibrium.

sigce

the process error on the relationship between cpue and exploitable biomass, set to 1e-08 when searching for an equilibrium.

r0

the unfished R0 level used by oneyearrec

b0

the unfished mature biomass B0, used in recruitment and depletion

exb0

the unfished exploitable biomass used in depletion

rdev

the recruitment deviates for each SAU for the given year

envyr

if notNULL contains the projeciton years in which perturbations in productiity will occur

envsurv

if notNULL contains post-larval survivorship for each year in envyr this is by popultion

envrec

if notNULL contains the proportion of stock recruitment in a given year that survives, by population

Value

a list containing a revised dynamics list

See also

Examples

print("Wait on new data")
#> [1] "Wait on new data"
#  zoneCC=zoneCP;inN=zoneDP$Nt[,year-1,,iter];popC=popC;year=year;
#  Ncl=Nclass;sauindex=sauindex;movem=movem;sigmar=sigmar;sigce=sigce;
#  r0=r0;b0=b0;exb0=exb0;rdev=-1;envyr=envyr;envsurv=survNt;envrec=proprec