oneyrgrowth one years' growth for a population and initial size

oneyrgrowth one years' growth for a given population and initial size. Used to determine the size after two year's growth as a measure of the expected bLML - biological Legal Minimum Length. A reflection of the two year rule in Tasmania. To get this the function should obviously be run twice.

oneyrgrowth(inpop, startsize = 2)

Arguments

inpop

the abpop to be grown forward

startsize

default = 2, but often set to the size at 50 percent maturity or the vector of midpts to determine the growth of Nt

Value

the expected mean size after one year's growth

Examples

if (FALSE) {
nblock <- 2
filename <- datafiletemplate(numblock=nblock,filename="block3.csv")
condDat <- readdatafile(filename)
const <- condDat$constants
blkpop <- condDat$blkpop
numpop <- condDat$numpop
popdefs <- definepops(nblock,blockI,const)
pop1 <- makeabpop(popdefs[1,],condDat$midpts,
                  condDat$ProjLML[popdefs[1,20]])
oneyrgrowth(pop1,startsize=70)
}