plotNt plots the size-composition for each SAU

plotNt accepts size-composition data (either the population Nt, of the catchN), and plots the replicate size-distributions for each SAU. It then puts the first year's size-distribution on top and if the medcol that defines the colour of hte mdeian of the replicates is different from 0 (not plotted) then it plots the median as well. A contrasting colour choice is 4 = Blue.

plotNt(Nt, year, glb, start = 3, medcol = 0)

Arguments

Nt

the 4 dimensional array of numbers-at-size Nclass,years,nsau,reps.

year

which year from dimension 2 should be plotted?

glb

the global constants object

start

from what size class should the maximum Y value be measured. The default = 3 (= 6mm)

medcol

colour of the median line; default=0 = not plotted

Value

invisibly the matrix of median values, Nclass x nsau

Examples

print("wait on suitable built in data-sets")
#> [1] "wait on suitable built in data-sets"