plothistce generates a plot of the historical cpue by SAU

plothistce generates a plot of historical cpue by SAU. It can plot as cpue or proportion of maximum cpue.

plothistce(rundir, condC, glb, proportion = FALSE, filen = "")

Arguments

rundir

the direcotry where all results are stored

condC

the R object contiaining the historical fishery data generated by the readctrlfile function

glb

the globol variables object

proportion

should the plot be of proportion relative to the maximum catch through each time-series or as raw tonnes x year, default=FALSE

filen

the name of the file used to save the plot, default="", which sends the plot to a separate window

Value

nothing but it does generate a plot either to a window or a file

Examples

print("wait on suitable data sets")
#> [1] "wait on suitable data sets"