plotzonedyn plots zonewide catch, mature biomass, harvestR and CPUE

plotzonedyn generates plots of the zone-wide total catches, the mature biomass, the harvest rate, and the CPUE across the years of projection.

plotzonedyn(
  rundir,
  scenes,
  zone,
  glb,
  console = TRUE,
  q90 = TRUE,
  polys = TRUE,
  intens = 127
)

Arguments

rundir

the directory in which all results are held for a scenario or comparison of scenarios

scenes

the names of the different scenarios being compared

zone

a list of the out$outzone objects from each chosen scenario

glb

one of the global objects from out. They should all have the same reps and years.

console

should the plot be sent to the console or saved for use in the web-page output? default=TRUE, set to FALSE to save it

q90

should the 90th quantiles be plotted? default=TRUE. If FALSE then the 95th quantiles are used.

polys

should transparent polygons be plotted = TRUE, or lines = FALSE

intens

if polys=TRUE then intens signifies the intensity of colour on a scale of 0 - 255. 127 is about 50 percent dense.

Value

nothing but it does generate a plot and saves it if console=FALSE

See also

Examples

print("wait on data sets")
#> [1] "wait on data sets"