plotzonesau generates a plot of the zone scale and sau scale

plotsonesau enables the zone catch or catch weighted zone cpue (from catchweightCE) to be compared with the sau catch or the sau cpue within a single plot. This has value for gaining insights into the fishery dynamics of a zone and how it responds to fishing pressure. NOTE WELL: Currently, this function is setup for Tasmania's set of 8 SAU. Assuming other jurisdictions have more or less then the layout of the plots would need to be altered. Until these are known this function will likely only work as intended with Tasmanian data!

plotzonesau(
  zonetot,
  saudat,
  saunames,
  label,
  labelsau,
  side = 3,
  sauscale = TRUE
)

Arguments

zonetot

a vector of either the total catch or zone summary cpue for a zone. Ideally, the vector should be named using the year in which each value was taken.

saudat

either the catch or the cpue for each SAU (year x sau), again this matrix should have both row and column names

saunames

a vector of nSAU names. Ideally very short ones to ensure they fit on each small plot

label

the y-axis label for the zone plot

labelsau

the y-axis label for each row of the SAU plots

side

should the SAU label be at the top of bottom of each plot. default = 3 (= top of plot). The main alternative would be 1 (the bottom) but 2 (left side) and 4 (right side) also work.

sauscale

should all SAU plots have the own y-axis scale = TRUE, the default. If you want each plot to have the same scale sauscale=FALSE.

Value

nothing but this does generate a plot.

See also

Examples

print("wait on suitable internal dats sets")
#> [1] "wait on suitable internal dats sets"