poptozone translates the zone_pop objects to a single zone object

poptozone combines the dynamic results for each variable so that results by population become results by zone. matureB, exploitB, catch, recruit, catchN, and Nt are simple summations of the totals for each population into their respective Zone The harvest rate would be end of year or beginning of year estimates derived from dividing the catch x zone by the exploitable biomass x zone. Similarly the deplsB and depleB are the end of year matureB and exploitB divided by their respective unfished estimated by Zone obtained using getvar(zoneC,"B0"). The cpue is a special case as combining across populations and sau requires the values to be catch weighted rather than giving them equal weight. So the cpue for each population is multiplied by the catch by population divided by the total catch and the result is summed. Thus, those contributing the greatest proportion of the catch contribute the most ot the cpue estiamte.

poptozone(inzone, NAS = NULL, glb, B0, ExB0)

Arguments

inzone

one of the zone dynamics objects containing replicates, made up of populations

NAS

the numbers-at-size 4D arrays from doprojection; default=NULL so it can be ignored during conditioning

glb

the object containing the global constants

B0

the sum of B0 across all populations, use getvar(zoneC,"B0")

ExB0

the sum of ExB0 across all populations use getvar(zoneC,"ExB0")

Value

a list of dynamics variables by zone

Examples

print("wait on suitable internal data sets ")
#> [1] "wait on suitable internal data sets "