prepareprojection high level function that sets up a projection

prepareprojection is a high level function that restructures zoneC by including the projected selectivity (in case the LML is set to change during the projections), and also selectivity x weight-at-size (for computational speed). It then generates a new zoneD with room for all replicates as well as the aspirational catches from each HS. It does this by converting the arrays of year x pop, to year x pop x replicate. It then uses the conditioned data in zoneDep to predict the first aspirational catches for the projections and conducts the initial replicate, thus starting the application of the HS. Finally, it adds recruitment variation to each of the replicates and keeps the last year of each iteration of the addrecvar function as the start of each replicate projection, with zeros in the catch, cpue, and cesau arrays. If the repeatable projections are wanted either set _randseedP_ = 0, whereupon the pseudorandom numbers will proceed from those started using the _randomseed_ set in the control file. Alternatively set a new seed into _randseedP_ (perhaps use _getseed()_), or set it to NA, which means it will merely run _set.seed()_, which will use the system time as a starting seed (hence each scenario run will differ).

prepareprojection(
  projC = projC,
  condC = condC,
  zoneC = zoneC,
  glb = glb,
  zoneDD = zoneDD,
  ctrl = ctrl,
  varyrs = varyrs,
  calcpopC = calcpopC,
  lastsigR = 0.3
)

Arguments

projC

the projection object from readctrlfile

condC

historical conditioning data

zoneC

the constant part of the zone

glb

the global variables

zoneDD

the zone after initial depletion through conditioning on the fishery

ctrl

the ctrl object for the scenario run

varyrs

how many years at the end to add recruitment variation

calcpopC

a function that takes the output from hcrfun and gernerates the actual catch per population expected in the current year.

lastsigR

recruitment variation for when it is applied for varyrs

Value

a list of the dynamic zone object as a list of arrays of projyrs x populations x replicates, plus the revised projC and revised zoneC

Examples

print("wait on data files")
#> [1] "wait on data files"