readctrlfile reads the rundir scenario's control file

readctrlfile checks rundir contains the required csv files including the named control file, which then contains the names of the region data file, and the population data file. The run stops if any are not present or are misnamed.

readctrlfile(rundir, infile = "control.csv", verbose = TRUE, fisreadfun = NULL)

Arguments

rundir

the directory in which all files relating to a particular run are to be held.

infile

default="control.csv", the filename of the control file present in rundir containing information regarding the run.

verbose

Should progress comments be printed to console, default=TRUE

fisreadfun

a function to read in the fissettings.csv file if a FIS is being used.

Value

the control list for the run

Examples

if (FALSE) {
# this has sinoce been modified and needs updating
rundir <- tempdir()
ctrlzonetemplate(rundir)
datafiletemplate(6,rundir,filename="zone1sau2pop6.csv")
ctrl <- readctrlfile(rundir)
print(ctrl)
}