readsaudatafile generates the constants matrix from sau data

readsaudatafile uses data described at the SAU level to make the constants file, which is then used to generate the popdefs matrix containing the specific values. This change from explicitly defining each population has been implemented to simplify the conditioning of each operating model.

readsaudatafile(rundir, infile, optpar = NULL)

Arguments

rundir

the directory in which the data file is to be found. This will usually be the rundir for the scenario run

infile

the name of the specific datafile used.

optpar

the optimum parameters from sizemod used to replace inputs for the main parameters estimated.

Value

a list of the constants matrix with values for each population and the original matrix of sau values from readsaudatafile

Examples

print("wait on suitable data sets")
#> [1] "wait on suitable data sets"
# rundir=rundir; infile=ctrl$datafile;optpar=opar