saudat is a matrix of constants read by readsaudatafile

saudat is a 32 x 2 matrix of values for biological variables relating to growth, weight-at-size, maturity-at-size, average unfished recruitment, selectivity, and other details, including the variation expected for each main variable. This information is used to define the populations within each SAU.

Format

A data.frame of biological and fishery parameters

DLMax

maximum growth increment

L50

size at half the maximum growth increment

L50inc

L95 - L50, size where growth increment only 5percent of max

Wtb

exponent of weight-at-size relationship

Wta

intercept of weight-at-size relationship

Examples

 data(saudat)
 round(saudat,5)
#>                           labels for each SAU
#> DLMax           28.60977             28.60977
#> sMaxDL           0.00000              0.00000
#> L50            130.00000            130.00000
#> sL50             0.00000              0.00000
#> L50inc          51.52389             51.52389
#> sL50inc          0.00000              0.00000
#> SigMax           3.37840              3.37840
#> sSigMax          0.00000              0.00000
#> LML            132.00000            132.00000
#> Wtb              3.16205              3.16205
#> sWtb             0.00000              0.00000
#> Wta              0.00006              0.00006
#> sWta             0.00000              0.00000
#> Me               0.15000              0.15000
#> sMe              0.00000              0.00000
#> AvRec      1764761.00000        1764761.00000
#> sAvRec           0.00000              0.00000
#> defsteep         0.70000              0.70000
#> sdefsteep        0.00000              0.00000
#> L50C           126.42220            126.42220
#> sL50C            0.00000              0.00000
#> deltaC          10.00000             10.00000
#> sdeltaC          0.00000              0.00000
#> MaxCEpars        0.45000              0.45000
#> sMaxCEpars       0.00000              0.00000
#> selL50p          0.00000              0.00000
#> selL95p          4.10480              4.10480
#> SaMa           -15.21940            -15.21940
#> L50Mat         112.23746            112.23746
#> sL50Mat          0.00000              0.00000
#> lambda           0.75000              0.75000
#> qest             0.77798              0.77798