sauribbon a ribbon plot the quantiles x scenes for the input variable

sauribbon is used to produce a ribbon plot for a given sau across a set of scenarios for a given variable (cpue, catch, harvest rate, etc) out of the dynamics object. After running do_comparison

sauribbon(
  rundir,
  scenes,
  sau,
  varqnts,
  glb,
  varname,
  console = TRUE,
  q90 = TRUE,
  intens = 127,
  addleg = "bottomright",
  defpar = TRUE
)

Arguments

rundir

the directory in which all results are held for a scenario or comparison of scenarios

scenes

the names of the different scenarios being compared

sau

a number from 1:nsau identifying which sau's dat to plot

varqnts

the qunatiles for each scenario as output by sauquantbyscene

glb

one of the global objects from out. They should all have the same reps and years.

varname

just the name of the variable being plotted, to make sure the figures all have the correct labelling

console

should the plot be sent to the console or saved for use in the web-page output? default=TRUE, set to FALSE to save it

q90

should the 90th quantiles be plotted? default=TRUE. If FALSE then the 95th quantiles are used.

intens

if polys=TRUE then intens signifies the intensity of colour on a scale of 0 - 255. 127 is about 50 percent dense.

addleg

add a legend or not. If not use "", if yes use any of topleft, topright, bottomleft, or bottomright, default = bottomright.

defpar

should the formatting of the plot use the internal version (the default setting) or, if set = FALSE use an external definition.

Value

nothing but it does generate a plot and saves it if console=FALSE

See also

Examples

if (FALSE) {
 # a possible example run where scenarios 1, 13, and 14 are compared.
   result <- do_comparison(rundir=rundir,postfixdir=postfixdir,outdir=outdir,
                           files=files,pickfiles=c(1,13,14),verbose=TRUE)
   out <- do_comp_outputs(result,projonly=TRUE)
   catch <- scenebyvar(dyn=out$dyn,byvar="catch",glb=out$glbc[[1]])
   sauribbon("",scenes=out$scenes,sau=8,varqnts=catqnts,glb=out$glbc[[1]],
   varname="Catch",console=TRUE,q90=TRUE,intens=100,addleg="bottomright")
}