scaleto1 scales an input vector of CPUE to a mean of one x avCE

scaleto1 scales a vector of CPUE to a mean of one or avCE. The use of a mean of one means that visual comparisons between different time-series becomes visually simplified. The avCE option could be used to scale the CPUE to the average geometric mean - so as to put it on the nominal scale

scaleto1(invect)

Arguments

invect

a vector of linear scale CPUE

Value

a vector of CPUE re-scaled to a mean of one

Examples

 ce <- c(0.4667187,1.2628564,0.8442146,0.9813531, 0.5554076,0.7426321)
 scaleto1(ce)
#> [1] 0.5770053 1.5612721 1.0437043 1.2132490 0.6866516 0.9181177