tasab is a matrix of abalone maturity-at-length data

tasab is a 715 x 4 matrix of maturity-at-length data for blacklip abalone (Haliotis rubra) from two sites along the Tasmanian west coast. All data was collected in February 1995, but details, such as site name, accurate location, statistical block, year, month, and other details have been omitted for brevity.

Format

A data.frame of maturity-at-length data

site

an identifier for the two different sites sampled

sex

I = immature, M = male, F = female

length

the shell length in mm

mature

was the animal mature = 1 or not = 0

Source

Thanks to the Institute of Marine and Antarctic Science, which is part of the University of Tasmania, and especially to Dr Craig Mundy, leader of the Abalone Group, for permission to use this data collected in February 1995.

Subjects

maturity ogives or logistic curves

Binomial likelihoods

Examples

 data(tasab)
 head(tasab,20)
#>     site sex length mature
#> 231    1   I    102      0
#> 233    1   I    102      0
#> 107    1   I    103      0
#> 242    1   I    105      0
#> 119    1   I    108      0
#> 4      1   F    110      1
#> 117    1   I    110      0
#> 109    1   I    111      0
#> 112    1   F    112      1
#> 64     1   F    113      1
#> 74     1   M    115      1
#> 204    1   F    115      1
#> 216    1   F    115      1
#> 21     1   F    116      1
#> 22     1   I    116      0
#> 123    1   F    116      1
#> 124    1   F    116      1
#> 27     1   F    118      1
#> 108    1   M    118      1
#> 113    1   F    118      1
 table(tasab$site,tasab$sex)
#>    
#>       F   I   M
#>   1 116  11 123
#>   2 207  85 173