testequil runs a zone for hyrs and determines stability

testequil runs a given zone for hyrs at the given harvest rate, and then tests that the last values of matureB, exploitB, recruitment, and spawning biomass depletion are the same as the first (to three decimal places). It reports this to the console if verbose=TRUE. This is used with a harvest rate of zero and no variation in recruitment when defining the equilibrium zone under the application of the movement matrix.

testequil(zoneC, zoneD, glb, inH = 0, verbose = TRUE)

Arguments

zoneC

the constants components of the simulated zone

zoneD

the dynamic components of the simulated zone

glb

the global variables

inH

a vector of numpop harvest rates

verbose

should results go to the console, default=TRUE

Value

the dynamics component with hyrs of dynamics

Examples

if (FALSE)   # modzoneC takes too long to run because of doproduction
 data(zone)
 zoneDe <- testequil(zoneC=zone$zoneC,zoneD=zone$zoneD,glb=zone$glb)
#> matureB Stable 
#> exploitB Stable 
#> recruitment varies 
#> spawning depletion Stable 
    #zoneC=zoneC; zoneD=zoneD; glb=glb; inH=0.0; verbose=TRUE